banking/bank-of-ai/pyproject.toml
[project]
name = "bank-of-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.97.0",
    "python-dotenv>=1.1.1",
]
